Added command pack/unpack unit tests (#144)
# Purpose
Added more unit tests for command pack/unpack functions (and the
pack/unpack utils)

# New Changes
- Moved command_pack module to the cdh directory in preparation for the
future directory revamp where all command stuff will be in a single
directory.
- Changed argument in all pack functions from size_t to uint32_t because
the unpack functions use uint32_t

# Testing
- CI

# Outstanding Changes
- N/A

+++ cdh/source/command_pack.c
@@ -8,7 +8,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-typedef void (*pack_func_t)(uint8_t*, size_t*, const cmd_msg_t*);
+typedef void (*pack_func_t)(uint8_t*, uint32_t*, const cmd_msg_t*);
 
 static const pack_func_t packFns[] = {
     [CMD_EXEC_OBC_RESET] = packExecObcResetCmdData,
@@ -23,7 +23,7 @@ static const pack_func_t packFns[] = {
 #define MAX_CMD_ID ((sizeof(packFns) / sizeof(pack_func_t)) - 1)
 
 // Pack the command message
-obc_error_code_t packCmdMsg(uint8_t* buffer, size_t* offset, const cmd_msg_t* cmdMsg, uint8_t* numPacked) {
+obc_error_code_t packCmdMsg(uint8_t* buffer, uint32_t* offset, const cmd_msg_t* cmdMsg, uint8_t* numPacked) {
   if (buffer == NULL || offset == NULL || cmdMsg == NULL) {
     return OBC_ERR_CODE_INVALID_ARG;
   }
@@ -49,31 +49,31 @@ obc_error_code_t packCmdMsg(uint8_t* buffer, size_t* offset, const cmd_msg_t* cm
 }
 
 // CMD_EXEC_OBC_RESET
-void packExecObcResetCmdData(uint8_t* buffer, size_t* offset, const cmd_msg_t* cmdMsg) {
+void packExecObcResetCmdData(uint8_t* buffer, uint32_t* offset, const cmd_msg_t* cmdMsg) {
   // No data to pack
 }
 
 // CMD_RTC_SYNC
-void packRtcSyncCmdData(uint8_t* buffer, size_t* offset, const cmd_msg_t* cmdMsg) {
+void packRtcSyncCmdData(uint8_t* buffer, uint32_t* offset, const cmd_msg_t* cmdMsg) {
   packUint32(cmdMsg->rtcSync.unixTime, buffer, offset);
 }
 
 // CMD_DOWNLINK_LOGS_NEXT_PASS
-void packDownlinkLogsNextPassCmdData(uint8_t* buffer, size_t* offset, const cmd_msg_t* cmdMsg) {
+void packDownlinkLogsNextPassCmdData(uint8_t* buffer, uint32_t* offset, const cmd_msg_t* cmdMsg) {
   packUint8((uint8_t)cmdMsg->downlinkLogsNextPass.logLevel, buffer, offset);
 }
 
 // CMD_MICRO_SD_FORMAT
-void packMicroSdFormat(uint8_t* buffer, size_t* offset, const cmd_msg_t* cmdMsg) {
+void packMicroSdFormat(uint8_t* buffer, uint32_t* offset, const cmd_msg_t* cmdMsg) {
   // No data to pack
 }
 
 // CMD_PING
-void packPingCmdData(uint8_t* buffer, size_t* offset, const cmd_msg_t* cmdMsg) {
+void packPingCmdData(uint8_t* buffer, uint32_t* offset, const cmd_msg_t* cmdMsg) {
   // No data to pack
 }
 
 // CMD_DOWNLINK_TELEM
-void packDownlinkTelemCmdData(uint8_t* buffer, size_t* offset, const cmd_msg_t* msg) {
+void packDownlinkTelemCmdData(uint8_t* buffer, uint32_t* offset, const cmd_msg_t* msg) {
   // No data to pack
 }
