+++ comms/source/encode_telemetry.c
@@ -317,7 +317,7 @@ static obc_error_code_t sendOrPackNextTelemetry(telemetry_data_t *singleTelem, p
   LOG_DEBUG("Sending telemetry: %u", singleTelem->id);
 
   uint8_t packedSingleTelem[MAX_TELEMETRY_DATA_SIZE];  // Holds a serialized version of the current piece of telemetry
-  size_t packedSingleTelemSize = 0;                    // Size of the packed single telemetry
+  uint32_t packedSingleTelemSize = 0;                  // Size of the packed single telemetry
 
   // Pack the single telemetry into a uint8_t array
   RETURN_IF_ERROR_CODE(packTelemetry(singleTelem, packedSingleTelem, sizeof(packedSingleTelem) / sizeof(uint8_t),
